requirements.txt
logzero
requests
python-dateutil
boto3
gsearch
feedgen
